requirements.txt
# BGAI v3.0.0 — Python Dependencies
# Install: pip install -r requirements.txt
# Python 3.10+ required

# ── Web Framework & API ─────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Database & ORM ──────────────────────────────────────────────────────────
sqlalchemy>=2.0.30

# ── Data Validation ─────────────────────────────────────────────────────────
pydantic[email]>=2.6.0
email-validator>=2.1.1

# ── Authentication & Security ────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# ── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# ── Visualization ────────────────────────────────────────────────────────────
plotly>=5.20.0
matplotlib>=3.8.0

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.42.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
watchdog>=4.0.0
